Tons of Overhaul Stuff
- Tons of attachments changes have been applied:
Made available ALL currently remade attachments with incomplete attachments available as temporary placeholders. This means that the following are all remade from scratch, or close to:
∟ Ammo Types
∟ Muzzle Attachments
∟ Optics
∟ Perks*
∟ Stocks
∟ Tacticals
∟ Underbarrels
∟ And more!
Additionally, all of the Assault Rifle attachments are also redone, excluding the SIG556 at the time of this being pushed.
This does mean that the majority of these attachments may be completely redesigned and will behave completely differently. Please provide feedback either on the Diamond Doves Discord, or on this Github page.

- Weapon changes have been applied:
- Galil AR: When equipped with the "MEPRO M21" optic, the optic is now placed further forward to replicate the early version of CS:GO.
- Galil SAR: Adjusted animations, making most of them faster; will keep adjusting over time.
- M4A1-S: Performed rework*, and reduced the spread multiplier when suppressed.
- M4A4: Performed rework*.
- SCAR-17: Performed rework*.
- SIG556: Performed partial rework**.

*Rework means the following is added or changed: Recoil, Spread, Aim Sway, CustomizePos, unique PeekPos, altered Sprint animation times and added RefillProgress to reloads.
**Partial rework means the following is added or changed: Spread, Aim Sway, CustomizePos, unique (albeit incomplete) PeekPos and altered Sprint animation times.